more undo/redo improvements
