+++ Catan10.csproj
@@ -119,6 +119,8 @@
     <Compile Include="Games\ExpansionCtrl.xaml.cs">
       <DependentUpon>ExpansionCtrl.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Logging\LogDefinitions.cs" />
+    <Compile Include="Logging\LogObjects.cs" />
     <Compile Include="MainPage\MainPageDependencyProperties.cs" />
     <Compile Include="MainPage\RoadRaceTracking.cs" />
     <Compile Include="MainPage\webSocket.cs" />
@@ -130,7 +132,7 @@
     <Compile Include="UI Controls\ResourceCardCtrl.xaml.cs">
       <DependentUpon>ResourceCardCtrl.xaml</DependentUpon>
     </Compile>
-    <Compile Include="MainPage\logging.cs" />
+    <Compile Include="Logging\logging.cs" />
     <Compile Include="ContentDialogs\NewGameDlg.xaml.cs">
       <DependentUpon>NewGameDlg.xaml</DependentUpon>
     </Compile>
