+++ Logging/LogDefinitions.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Runtime.CompilerServices;
+using System.Text;
+using System.Threading.Tasks;
+using Windows.Storage;
+using Windows.Storage.Streams;
+
+namespace Catan10
+{
+    public enum LogState
+    {
+        Normal,
+        Replay,
+        Undo
+    }
+
+    public interface ILogParserHelper
+    {
+        TileCtrl GetTile(int tileIndex, int gameIndex);
+        RoadCtrl GetRoad(int roadIndex, int gameIndex);
+        BuildingCtrl GetBuilding(int buildingIndex, int gameIndex);
+        PlayerData GetPlayerData(int playerIndex);
+    }
+
+    public interface ILog
+    {
+        Task AddLogEntry(PlayerData player, GameState state, CatanAction action, bool stopProcessingUndo, LogType logType = LogType.Normal, int number = -1, object tag = null, [CallerFilePath] string filePath = "", [CallerMemberName] string name = "", [CallerLineNumber] int lineNumber = 0);
+        void PostLogEntry(PlayerData player, GameState state, CatanAction action, bool stopProcessingUndo, LogType logType = LogType.Normal, int number = -1, object tag = null, [CallerFilePath] string filePath = "", [CallerMemberName] string name = "", [CallerLineNumber] int lineNumber = 0);
+    }
+
+    public enum LogType { Normal, Undo, Replay, DoNotLog, DoNotUndo };
+
+
+
+    public delegate void RedoPossibleHandler(bool redo);
+}
